# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_minisource_log_internal_models.LogRetention:
    properties:
      archive_enabled:
        type: boolean
      archive_path:
        type: string
      created_at:
        type: string
      id:
        type: string
      max_size_gb:
        type: integer
      retention_days:
        type: integer
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handler.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
host: localhost:5002
info:
  contact: {}
  description: Centralized logging service for Minisource
  title: Log Service API
  version: "1.0"
paths:
...
  /retention:
    get:
      description: Lists all retention policies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_minisource_log_internal_models.LogRetention'
            type: array
      summary: List retention policies
      tags:
      - retention
    post:
      consumes:
      - application/json
      description: Creates a new retention policy for a tenant
      parameters:
      - description: Retention Policy
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/github_com_minisource_log_internal_models.LogRetention'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_minisource_log_internal_models.LogRetention'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Create retention policy
      tags:
      - retention
  /retention/{id}:
    delete:
      description: Deletes a retention policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Delete retention policy
      tags:
      - retention
    put:
      consumes:
      - application/json
      description: Updates an existing retention policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      - description: Retention Policy
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/github_com_minisource_log_internal_models.LogRetention'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_minisource_log_internal_models.LogRetention'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Update retention policy
      tags:
      - retention
  /retention/tenant/{tenant_id}:
    get:
      description: Retrieves retention policy for a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_minisource_log_internal_models.LogRetention'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Get retention policy
      tags:
      - retention
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"